requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
dynmap-timemachine==0.9.5
future==0.18.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
Pillow==7.1.2
PySocks==1.7.1
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.9
